requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
anaconda-anon-usage @ file:///opt/miniconda3/conda-bld/anaconda-anon-usage_1764636648062/work
anaconda-auth @ file:///home/task_176583888028164/conda-bld/anaconda-cloud-auth-split_1765839818582/work
anaconda-cli-base @ file:///home/task_176488869080098/conda-bld/anaconda-cli-base_1764888740587/work
annotated-types @ file:///home/task_176174487847081/conda-bld/annotated-types_1761744928471/work
anyio==4.12.1
archspec @ file:///home/task_175812491784513/conda-bld/archspec_1758124989039/work
arrow==1.4.0
asgiref==3.11.1
attrs==26.1.0
binaryornot==0.6.0
blinker==1.9.0
boltons @ file:///croot/boltons_1751383775419/work
brotlicffi @ file:///home/task_176496126124857/conda-bld/brotlicffi_1764961326114/work
certifi @ file:///home/task_176313455153976/conda-bld/certifi_1763135329031/work/certifi
cffi @ file:///home/task_176183214835616/conda-bld/cffi_1761832774688/work
charset-normalizer @ file:///home/task_176174481524083/conda-bld/charset-normalizer_1761744828711/work
click @ file:///home/task_176433224228540/conda-bld/click_1764332256338/work
comfy-cli==1.6.0
conda @ file:///home/task_176556791054345/conda-bld/conda_1765567979493/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1755883788794/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1755123313635/work
conda-content-trust @ file:///work/perseverance-python-buildout/croot/conda-content-trust_1728487058234/work
conda-libmamba-solver @ file:///C:/miniconda3/conda-bld/conda-libmamba-solver_1764245615235/work/src
conda-package-handling @ file:///home/task_176236637997838/conda-bld/conda-package-handling_1762366405159/work
conda_package_streaming @ file:///home/task_176236152622180/conda-bld/conda-package-streaming_1762361633858/work
cookiecutter==2.7.1
cryptography @ file:///home/task_176193173929288/conda-bld/cryptography-split_1761931963149/work
distro @ file:///work/perseverance-python-buildout/croot/distro_1728396110052/work
et_xmlfile==2.0.0
filelock==3.25.0
Flask==3.1.3
frozendict @ file:///home/task_176175049677026/conda-bld/frozendict_1761750665451/work
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
gunicorn==25.1.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna @ file:///home/task_176191175488988/conda-bld/idna_1761911981595/work
itsdangerous==2.2.0
jaraco.classes @ file:///croot/jaraco.classes_1755516327832/work
jaraco.context @ file:///io/perseverance-python-buildout/croot/jaraco.context_1731706750665/work
jaraco.functools @ file:///croot/jaraco.functools_1740408111753/work
jeepney @ file:///tmp/build/80754af9/jeepney_1627537048313/work
Jinja2==3.1.6
json-logic==0.7.0a0
jsonpatch @ file:///work/perseverance-python-buildout/croot/jsonpatch_1728399595941/work
jsonpointer @ file:///croot/jsonpointer_1753788433671/work
keyring @ file:///home/task_176363713576885/conda-bld/keyring_1763637184857/work
libmambapy @ file:///home/task_176311149180614/conda-bld/mamba-split_1763111570862/work/libmambapy
markdown-it-py @ file:///croot/markdown-it-py_1756299186942/work
MarkupSafe==3.0.3
mdurl @ file:///home/task_175855216352428/conda-bld/mdurl_1758552176751/work
menuinst @ file:///home/task_176538230483207/conda-bld/menuinst_1765382366331/work
mixpanel==5.1.0
more-itertools @ file:///home/task_176112148319406/conda-bld/more-itertools_1761121494880/work
msgpack @ file:///croot/msgpack-python_1750958623279/work
multidict==6.7.1
numpy==2.4.3
openpyxl==3.1.5
packaging @ file:///home/task_176104878999627/conda-bld/packaging_1761049078606/work
pandas==3.0.1
pathspec==1.0.4
pkce @ file:///work/perseverance-python-buildout/croot/pkce_1728388093137/work
platformdirs @ file:///home/task_176235602006708/conda-bld/platformdirs_1762356487962/work
pluggy @ file:///croot/pluggy_1733169602837/work
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///home/task_175749564936714/conda-bld/pycparser_1757496040682/work
pydantic @ file:///home/task_176408303339170/conda-bld/pydantic_1764083529211/work
pydantic-settings @ file:///home/task_176416496069032/conda-bld/pydantic-settings_1764165165402/work
pydantic_core @ file:///home/task_176400924785639/conda-bld/pydantic-core_1764009747660/work
Pygments @ file:///home/task_176243137807904/conda-bld/pygments_1762431407645/work
PyJWT @ file:///home/task_176433208417911/conda-bld/pyjwt_1764332098101/work
PySocks @ file:///home/task_176175235884818/conda-bld/pysocks_1761753009495/work
python-dateutil==2.9.0.post0
python-dotenv @ file:///croot/python-dotenv_1745610241282/work
python-slugify==8.0.4
PyYAML==6.0.3
questionary==2.1.1
readchar @ file:///home/task_176061336965958/conda-bld/readchar_1760613437376/work
requests @ file:///home/task_176235918788486/conda-bld/requests_1762359591632/work
rich @ file:///home/task_176037557612320/conda-bld/rich_1760375632000/work
ruamel.yaml @ file:///home/task_176253595903435/conda-bld/ruamel.yaml_1762535972567/work
ruamel.yaml.clib @ file:///home/task_176253026438192/conda-bld/ruamel.yaml.clib_1762530278005/work
ruff==0.15.6
SecretStorage @ file:///home/task_175793143758247/conda-bld/secretstorage_1757931467244/work
semver @ file:///home/task_176190313959014/conda-bld/semver_1761903196627/work
setuptools==80.9.0
shellingham @ file:///home/task_176191203629823/conda-bld/shellingham_1761912204430/work
six==1.17.0
smmap==5.0.3
text-unidecode==1.3
tomli @ file:///croot/tomli_1753774587605/work
tomlkit==0.14.0
tqdm @ file:///home/task_176286304944547/conda-bld/tqdm_1762863349742/work
truststore @ file:///home/task_176252074050774/conda-bld/truststore_1762520903993/work
twilio==9.10.3
typer==0.20.0
typer-slim==0.20.0
typing-inspection @ file:///home/task_176061348700415/conda-bld/typing-inspection_1760614166875/work
typing_extensions @ file:///croot/typing_extensions_1756280817316/work
tzdata==2025.3
urllib3 @ file:///home/task_176539908095521/conda-bld/urllib3_1765399198587/work
uv==0.10.10
wcwidth==0.6.0
websocket-client==1.9.0
Werkzeug==3.1.6
wheel==0.45.1
xlrd==2.0.2
yarl==1.23.0
zstandard @ file:///home/task_175818846949212/conda-bld/zstandard_1758189073015/work
